backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6
pydantic==2.10.4
pydantic-settings==2.6.1
boto3==1.35.63
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7
httpx==0.27.2
openai==1.58.1
